+++ src/components/InstagramEmbed.tsx
@@ -1,0 +1,131 @@
+/**
+ * Instagram Embed Component
+ * 
+ * Shows Instagram posts as clickable cards that link to the actual posts
+ * since direct media embedding has CORS restrictions
+ */
+
+import React from 'react';
+import { InstagramPost } from '@/services/instagram';
+
+interface InstagramEmbedProps {
+  post: InstagramPost;
+  className?: string;
+}
+
+export const InstagramEmbed: React.FC<InstagramEmbedProps> = ({ post, className = '' }) => {
+  const formatDate = (timestamp: string): string => {
+    try {
+      const date = new Date(timestamp);
+      return date.toLocaleDateString('en-AU', { 
+        month: 'short', 
+        day: 'numeric' 
+      });
+    } catch {
+      return '';
+    }
+  };
+
+  const truncateCaption = (caption: string | undefined, maxLength: number = 100): string => {
+    if (!caption) return '';
+    if (caption.length <= maxLength) return caption;
+    return caption.substring(0, maxLength).trim() + '...';
+  };
+
+  return (
+    <a 
+      href={post.permalink}
+      target="_blank"
+      rel="noopener noreferrer"
+      className={`block bg-white rounded-xl shadow-sm hover:shadow-lg transition-all duration-300 overflow-hidden group ${className}`}
+    >
+      {/* Instagram Post Preview Card */}
+      <div className="p-6">
+        {/* Header */}
+        <div className="flex items-center justify-between mb-4">
+          <div className="flex items-center space-x-3">
+            {/* Instagram Icon */}
+            <div className="w-10 h-10 bg-gradient-to-br from-purple-500 via-pink-500 to-orange-400 rounded-lg flex items-center justify-center">
+              <svg className="w-6 h-6 text-white" fill="currentColor" viewBox="0 0 24 24">
+                <path d="M12 2.163c3.204 0 3.584.012 4.85.07 3.252.148 4.771 1.691 4.919 4.919.058 1.265.069 1.645.069 4.849 0 3.205-.012 3.584-.069 4.849-.149 3.225-1.664 4.771-4.919 4.919-1.266.058-1.644.07-4.85.07-3.204 0-3.584-.012-4.849-.07-3.26-.149-4.771-1.699-4.919-4.92-.058-1.265-.07-1.644-.07-4.849 0-3.204.013-3.583.07-4.849.149-3.227 1.664-4.771 4.919-4.919 1.266-.057 1.645-.069 4.849-.069zm0-2.163c-3.259 0-3.667.014-4.947.072-4.358.2-6.78 2.618-6.98 6.98-.059 1.281-.073 1.689-.073 4.948 0 3.259.014 3.668.072 4.948.2 4.358 2.618 6.78 6.98 6.98 1.281.058 1.689.072 4.948.072 3.259 0 3.668-.014 4.948-.072 4.354-.2 6.782-2.618 6.979-6.98.059-1.28.073-1.689.073-4.948 0-3.259-.014-3.667-.072-4.947-.196-4.354-2.617-6.78-6.979-6.98-1.281-.059-1.69-.073-4.949-.073zm0 5.838c-3.403 0-6.162 2.759-6.162 6.162s2.759 6.163 6.162 6.163 6.162-2.759 6.162-6.163c0-3.403-2.759-6.162-6.162-6.162zm0 10.162c-2.209 0-4-1.79-4-4 0-2.209 1.791-4 4-4s4 1.791 4 4c0 2.21-1.791 4-4 4zm6.406-11.845c-.796 0-1.441.645-1.441 1.44s.645 1.44 1.441 1.44c.795 0 1.439-.645 1.439-1.44s-.644-1.44-1.439-1.44z"/>
+              </svg>
+            </div>
+            <div>
+              <p className="font-semibold text-gray-900">@{post.username}</p>
+              <p className="text-sm text-gray-500">{formatDate(post.timestamp)}</p>
+            </div>
+          </div>
+          
+          {/* Post Type Badge */}
+          <div className="flex items-center space-x-1 bg-gray-100 px-2 py-1 rounded-full">
+            {post.media_type === 'VIDEO' ? (
+              <>
+                <svg className="w-4 h-4 text-gray-600" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                  <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M14.828 14.828a4 4 0 01-5.656 0M9 10h1.5a1.5 1.5 0 011.5 1.5v1a1.5 1.5 0 01-1.5 1.5H9m4.5-5H15a1.5 1.5 0 011.5 1.5v1a1.5 1.5 0 01-1.5 1.5h-1.5m-4-3v8m4-8v8" />
+                </svg>
+                <span className="text-xs text-gray-600">Reel</span>
+              </>
+            ) : (
+              <>
+                <svg className="w-4 h-4 text-gray-600" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                  <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M4 16l4.586-4.586a2 2 0 012.828 0L16 16m-2-2l1.586-1.586a2 2 0 012.828 0L20 14m-6-6h.01M6 20h12a2 2 0 002-2V6a2 2 0 00-2-2H6a2 2 0 00-2 2v12a2 2 0 002 2z" />
+                </svg>
+                <span className="text-xs text-gray-600">Post</span>
+              </>
+            )}
+          </div>
+        </div>
+
+        {/* Content Preview */}
+        <div className="space-y-3">
+          {/* Caption */}
+          <p className="text-gray-800 leading-relaxed">
+            {truncateCaption(post.caption)}
+          </p>
+
+          {/* Tags */}
+          {post.tags && post.tags.length > 0 && (
+            <div className="flex flex-wrap gap-1">
+              {post.tags.slice(0, 3).map((tag) => (
+                <span 
+                  key={tag}
+                  className="inline-block bg-blue-50 text-blue-600 text-xs px-2 py-1 rounded-full"
+                >
+                  #{tag}
+                </span>
+              ))}
+            </div>
+          )}
+
+          {/* Venue/Event Info */}
+          {(post.venue || post.event_type) && (
+            <div className="flex items-center text-sm text-gray-500">
+              <svg className="w-4 h-4 mr-1" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M17.657 16.657L13.414 20.9a1.998 1.998 0 01-2.827 0l-4.244-4.243a8 8 0 1111.314 0z" />
+                <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M15 11a3 3 0 11-6 0 3 3 0 016 0z" />
+              </svg>
+              {post.venue} {post.venue && post.event_type && '•'} {post.event_type}
+            </div>
+          )}
+        </div>
+
+        {/* Click to View CTA */}
+        <div className="mt-4 pt-3 border-t border-gray-100">
+          <div className="flex items-center justify-between">
+            <span className="text-sm text-gray-500">View on Instagram</span>
+            <svg 
+              className="w-5 h-5 text-gray-400 group-hover:text-blue-500 transform group-hover:translate-x-1 transition-all duration-200" 
+              fill="none" 
+              stroke="currentColor" 
+              viewBox="0 0 24 24"
+            >
+              <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M10 6H6a2 2 0 00-2 2v10a2 2 0 002 2h10a2 2 0 002-2v-4M14 4h6m0 0v6m0-6L10 14" />
+            </svg>
+          </div>
+        </div>
+      </div>
+    </a>
+  );
+};
+
+export default InstagramEmbed;
