+++ src/data/instagram-posts.json
@@ -3,7 +3,7 @@
     {
       "id": "reel-DNLMO67T4G8",
       "caption": "Latest musical performance - bringing the energy! 🎸🔥 Follow for more live music content! #livemusic #performance #music",
-      "media_url": "https://www.instagram.com/reel/DNLMO67T4G8/media/?size=l",
+      "media_url": "/images/instagram/performance-reel-1.webp",
       "media_type": "VIDEO",
       "permalink": "https://www.instagram.com/reel/DNLMO67T4G8/",
       "timestamp": "2024-12-01T18:30:00Z",
@@ -16,7 +16,7 @@
     {
       "id": "reel-DLaYVu4O7DS",
       "caption": "Musical moments that matter 🎵 Creating unforgettable experiences through music! #music #performance #live",
-      "media_url": "https://www.instagram.com/reel/DLaYVu4O7DS/media/?size=l",
+      "media_url": "/images/instagram/performance-reel-2.webp",
       "media_type": "VIDEO",
       "permalink": "https://www.instagram.com/reel/DLaYVu4O7DS/",
       "timestamp": "2024-11-20T20:15:00Z",
@@ -29,7 +29,7 @@
     {
       "id": "reel-DMQROZoOQYz",
       "caption": "Behind the music - every note tells a story 🎹✨ #piano #music #storytelling #performance",
-      "media_url": "https://www.instagram.com/reel/DMQROZoOQYz/media/?size=l",
+      "media_url": "/images/instagram/performance-reel-3.webp",
       "media_type": "VIDEO",
       "permalink": "https://www.instagram.com/reel/DMQROZoOQYz/",
       "timestamp": "2024-11-15T16:45:00Z",
@@ -42,7 +42,7 @@
     {
       "id": "post-DLR4m-PSA08",
       "caption": "Sharing the passion for music through performance and teaching 🎼 Every moment on stage is a gift! #musician #livemusic #passion",
-      "media_url": "https://www.instagram.com/p/DLR4m-PSA08/media/?size=l",
+      "media_url": "/images/instagram/performance-post-1.webp",
       "media_type": "IMAGE",
       "permalink": "https://www.instagram.com/p/DLR4m-PSA08/",
       "timestamp": "2024-11-10T19:20:00Z",
